serviceinvoke/circuit-breaker/src/main/java/com/turboic/cloud/vo/App.java
package com.turboic.cloud.vo;

import java.io.Serializable;
import java.util.Date;

/**
 * @author liebe
 */
public class App implements Serializable {
    private String name;
    private String version;
    private String description;
    private Date date;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getVersion() {
        return version;
    }

    public void setVersion(String version) {
        this.version = version;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Date getDate() {
        return date;
    }

    public void setDate(Date date) {
        this.date = date;
    }
}
